- Added Velich's tests. - Moved Zvi's tests to their own directory. - Simulator and tester are almost ready. - Changed yield in barrier_test to be a sleep, to be more platform independent.

+++ OperSystems/tests/ex3/zvi/zvi_test.c
@@ -89,7 +89,7 @@ void mpsend(void* v) {
 	}
 
 	mp_barrier(con,senders_barrier);								 /* wait for all sending threads to finish */
-	if (tid == final_message_sender) {								 /* only one of the sending threads will issue a final message */
+	if (tid == (int)final_message_sender) {					 /* only one of the sending threads will issue a final message */
 		printf("@ Sender %d sending FINAL BROADCAST...\n",tid);
     fflush(stdout);
 		retval = mp_broadcast(con,final,strlen(final)+1,0);
